refactor(agents): split direct compaction pipeline
